move most functions into their own package
